RPC client refactoring to divide responsibilities
